feat: use new BLE flow components in BLE screen
