+++ apps/ledger-live-mobile/src/screens/BleDevicePairingFlow/index.tsx
@@ -1,19 +1,15 @@
-import React, { useCallback } from "react";
-import { useDispatch } from "react-redux";
+import { Device, DeviceModelId } from "@ledgerhq/types-devices";
+import React, { useCallback, useEffect } from "react";
 import { has as hasFromPath, set as setFromPath } from "lodash";
-import { Device } from "@ledgerhq/live-common/hw/actions/types";
-import { DeviceModelId } from "@ledgerhq/types-devices";
-import RequiresBLE from "../../components/RequiresBLE";
-import { BleDevicesScanning } from "./BleDevicesScanning";
-import { BleDevicePairing } from "./BleDevicePairing";
-import { addKnownDevice } from "../../actions/ble";
+import { BackHandler } from "react-native";
 import { NavigatorName, ScreenName } from "../../const";
-import { useResetOnNavigationFocusState } from "../../helpers/useResetOnNavigationFocusState";
-import { BaseNavigatorStackParamList } from "../../components/RootNavigator/types/BaseNavigator";
+import { useIncrementOnNavigationFocusState } from "../../helpers/useIncrementOnNavigationFocusState";
+import BleDevicePairingFlowComponent from "../../components/BleDevicePairingFlow/index";
 import {
   RootComposite,
   StackNavigatorProps,
 } from "../../components/RootNavigator/types/helpers";
+import { BaseNavigatorStackParamList } from "../../components/RootNavigator/types/BaseNavigator";
 
 export type Props = RootComposite<
   StackNavigatorProps<
@@ -46,12 +42,11 @@ const defaultNavigationParams = {
   },
 };
 
-// A "done" state to avoid having the BLE scanning on the device that we just paired
-// and to which messages are going to be exchanged via BLE
-type PairingFlowStep = "scanning" | "pairing" | "done";
-
 /**
  * Screen handling the BLE flow with a scanning step and a pairing step
+ *
+ * Note: this screen will disappear in the future, but still necessary for now.
+ *
  * @param navigation react-navigation navigation object
  * @param route react-navigation route object. The route params are:
  * - filterByDeviceModelId: (optional, default to none) a device model id to filter on
@@ -83,8 +78,6 @@ type PairingFlowStep = "scanning" | "pairing" | "done";
  * @returns a JSX component
  */
 export const BleDevicePairingFlow = ({ navigation, route }: Props) => {
-  const dispatchRedux = useDispatch();
-
   const params = route?.params || defaultNavigationParams;
 
   const {
@@ -100,44 +93,11 @@ export const BleDevicePairingFlow = ({ navigation, route }: Props) => {
     navigationType = "navigate",
   } = onSuccessNavigateToConfig;
 
-  // Resets when the navigation goes back to this screen
-  const [pairingFlowStep, setPairingFlowStep] = useResetOnNavigationFocusState<
-    PairingFlowStep,
-    Props["navigation"]
-  >(navigation, "scanning");
-  // Resets when the navigation goes back to this screen
-  const [deviceToPair, setDeviceToPair] = useResetOnNavigationFocusState<
-    Device | null,
-    Props["navigation"]
-  >(navigation, null);
+  const keyToReset =
+    useIncrementOnNavigationFocusState<Props["navigation"]>(navigation);
 
-  const onDeviceSelect = useCallback(
-    (item: Device) => {
-      const deviceToPair = {
-        deviceId: item.deviceId,
-        deviceName: item.deviceName,
-        modelId: item.modelId,
-        wired: false,
-      };
-
-      setDeviceToPair(deviceToPair);
-      setPairingFlowStep("pairing");
-    },
-    [setDeviceToPair, setPairingFlowStep],
-  );
-
-  const onPaired = useCallback(
+  const onPairingSuccess = useCallback(
     (device: Device) => {
-      if (onSuccessAddToKnownDevices) {
-        dispatchRedux(
-          addKnownDevice({
-            id: device.deviceId,
-            name: device.deviceName ?? device.modelId,
-            modelId: device.modelId,
-          }),
-        );
-      }
-
       const hasDeviceParam = hasFromPath(navigateInput, pathToDeviceParam);
       if (hasDeviceParam) {
         setFromPath(navigateInput, pathToDeviceParam, device);
@@ -149,10 +109,6 @@ export const BleDevicePairingFlow = ({ navigation, route }: Props) => {
         );
       }
 
-      // Before navigating, to never come back a the successful pairing but to the scanning part
-      setDeviceToPair(null);
-      setPairingFlowStep("done");
-
       const params = navigateInput.params
         ? {
             ...navigateInput.params,
@@ -161,7 +117,8 @@ export const BleDevicePairingFlow = ({ navigation, route }: Props) => {
 
       if (navigationType === "push") {
         // @ts-expect-error this seems complicated to type properly
-        // the typings for react-navigation cannot reconciliate screens having "undefined" params and object params
+        // the typings for react-navigation cannot reconciliate screens having "undefined" params and object params.
+        // And this will disappear in a future screen.
         navigation.push(navigateInput.name, params);
       } else if (navigationType === "replace") {
         // @ts-expect-error this seems complicated to type properly
@@ -171,38 +128,57 @@ export const BleDevicePairingFlow = ({ navigation, route }: Props) => {
         navigation.navigate(navigateInput.name, params);
       }
     },
-    [
-      dispatchRedux,
-      navigateInput,
-      navigation,
-      navigationType,
-      onSuccessAddToKnownDevices,
-      pathToDeviceParam,
-      setDeviceToPair,
-      setPairingFlowStep,
-    ],
+    [navigateInput, navigation, navigationType, pathToDeviceParam],
   );
 
-  const onRetryPairingFlow = useCallback(() => {
-    setDeviceToPair(null);
-    setPairingFlowStep("scanning");
-  }, [setDeviceToPair, setPairingFlowStep]);
+  const handleGoBackFromScanning = useCallback(() => {
+    const routes = navigation.getState().routes;
+
+    const isNavigationFromDeeplink =
+      routes[routes.length - 1]?.params === undefined;
+
+    if (!isNavigationFromDeeplink) {
+      navigation.goBack();
+    } else {
+      navigation.reset({
+        index: 0,
+        routes: [
+          {
+            // @ts-expect-error is fixed in a future screen
+            name: NavigatorName.BaseOnboarding,
+            state: {
+              routes: [
+                {
+                  name: ScreenName.OnboardingWelcome,
+                },
+              ],
+            },
+          },
+        ],
+      });
+    }
+  }, [navigation]);
+
+  // Handles back button, necessary when the user comes from the deep link
+  useEffect(() => {
+    const listener = BackHandler.addEventListener("hardwareBackPress", () => {
+      handleGoBackFromScanning();
+      return true;
+    });
+
+    return () => listener.remove();
+  }, [handleGoBackFromScanning]);
+
+  console.log(`🕵️ BleDevicePairingFlowScreen: keyToReset: ${keyToReset}`);
 
   return (
-    <RequiresBLE>
-      {pairingFlowStep === "pairing" && deviceToPair !== null ? (
-        <BleDevicePairing
-          deviceToPair={deviceToPair}
-          onPaired={onPaired}
-          onRetry={onRetryPairingFlow}
-        />
-      ) : pairingFlowStep === "scanning" ? (
-        <BleDevicesScanning
-          filterByDeviceModelId={filterByDeviceModelId}
-          areKnownDevicesDisplayed={areKnownDevicesDisplayed}
-          onDeviceSelect={onDeviceSelect}
-        />
-      ) : null}
-    </RequiresBLE>
+    <BleDevicePairingFlowComponent
+      key={keyToReset}
+      filterByDeviceModelId={filterByDeviceModelId}
+      areKnownDevicesDisplayed={areKnownDevicesDisplayed}
+      onGoBackFromScanning={handleGoBackFromScanning}
+      onPairingSuccess={onPairingSuccess}
+      onPairingSuccessAddToKnownDevices={onSuccessAddToKnownDevices}
+    />
   );
 };
